Fix de usuarios y de cómo se ven por los porfesores

+++ ProyectoFinal/Nutriapp/database/migrations/2026_04_23_090754_create_alimentos_table.php
@@ -15,7 +15,7 @@ return new class extends Migration
             $table->id('id');
             $table->foreignId('user_id')->constrained('usuarios')->cascadeOnDelete();
             $table->string('alimento', 255);
-            $table->unsignedInteger('pc');
+            $table->float('pc');
             $table->float('e_100',10,2);
             $table->float('prot_100',10,2);
             $table->float('grasa_100',10,2);
